requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.19.1
fsspec==2025.9.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.5.2
scipy==1.16.1
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.67.1
transformers==4.44.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
